Move base into subdir
