+++ pt_os_web_portal/miniscreen_onboarding/menu_pages/_unused/info.py
@@ -1,8 +1,8 @@
 from pitop.common.pt_os import get_pitopOS_info
 
 from ..menus import Menus
-from ._base import MenuPageBase
 from .attr.margins import FIRST_LINE_Y, INFO_PAGE_MARGIN_X, SECOND_LINE_Y, THIRD_LINE_Y
+from .base._base import MenuPageBase
 from .render.helpers import draw_text
