+++ pt_os_web_portal/miniscreen_onboarding/menu_pages/base/_base.py
