+++ pt_os_web_portal/miniscreen_onboarding/menu_pages/carry_on.py
@@ -4,8 +4,8 @@ from time import sleep
 
 from ...backend.helpers.extras import started_onboarding_breadcrumb
 from ..menus import Menus
-from ._title_base import TitleMenuPage
 from .attr.margins import FIRST_LINE_Y, SECOND_LINE_Y, THIRD_LINE_Y
+from .base._title_base import TitleMenuPage
 from .render.helpers import draw_text
